+++ plugin-flex-ts-template-v2/src/feature-library/pause-recording/custom-components/PauseStatusPanel/PauseStatusPanel.tsx
@@ -1,4 +1,4 @@
-import { TaskHelper, ITask, Template, templates, Manager } from '@twilio/flex-ui';
+import { TaskHelper, ITask, Template, templates } from '@twilio/flex-ui';
 import React, { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { Text } from '@twilio-paste/core/text';
@@ -12,17 +12,8 @@ export interface OwnProps {
   task?: ITask;
 }
 
-const ALLOWED_LOCATIONS = ['us', 'uk'];
 const PauseStatusPanel = (props: OwnProps) => {
   const [paused, setPaused] = useState(false);
-
-  const manager = Manager.getInstance();
-  const workerLocation = manager.workerClient?.attributes?.location;
-  const showBannerBasedOnWorkerLocation = ALLOWED_LOCATIONS.includes(workerLocation);
-
-  console.log('Manager Details in Helper file', manager);
-  console.log('workerLocation Details in Helper file', workerLocation);
-  console.log('showBannerBasedOnWorkerLocation Details in Helper file', showBannerBasedOnWorkerLocation);
 
   const { pausedRecordings } = useSelector(
     (state: AppState) => state[reduxNamespace].pauseRecording as PauseRecordingState,
@@ -56,7 +47,7 @@ const PauseStatusPanel = (props: OwnProps) => {
 
   return (
     <>
-      {showBannerBasedOnWorkerLocation && paused && (
+      {paused && (
         <Text as="p" textAlign="center" fontWeight="fontWeightBold" padding="space50">
           <Template source={templates[StringTemplates.RECORDING_PAUSED_LABEL]} />
         </Text>
